fixup! feat(integration): test ADP dynamic config propagation
Address Toby's review: rename assertions->procedure and type->assertion
in YAML schema; extract CoreAgentConfigSetAction to its own file.
