+++ test/integration/cases/adp-logging-default-path/config.yaml
@@ -15,13 +15,13 @@ container:
   exposed_ports:
     - "58125/udp"
 
-assertions:
+procedure:
   # Wait for ADP to receive its initial configuration from the Core Agent and reload logging.
-  - type: log_contains
+  - assertion: log_contains
     pattern: "Initial configuration received."
     timeout: 30s
   # The default per-subagent log file path is used.
-  - type: file_contains
+  - assertion: file_contains
     path: "/var/log/datadog/agent-data-plane.log"
     pattern: "DATAPLANE"
     timeout: 30s
