+++ test/integration/cases/adp-logging-windows-default-path/config.yaml
@@ -15,11 +15,11 @@ container:
   exposed_ports:
     - "58125/udp"
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: "Initial configuration received."
     timeout: 60s
-  - type: file_contains
+  - assertion: file_contains
     path: "C:\\ProgramData\\Datadog\\logs\\agent-data-plane.log"
     pattern: "DATAPLANE"
     timeout: 60s
