+++ test/integration/cases/adp-logging-windows-ignores-core-agent-log-file/config.yaml
@@ -17,15 +17,15 @@ container:
   exposed_ports:
     - "58125/udp"
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: "Initial configuration received."
     timeout: 60s
-  - type: file_contains
+  - assertion: file_contains
     path: "C:\\adp\\adp-ignores-core-agent-log-file.log"
     pattern: "DATAPLANE"
     timeout: 60s
-  - type: file_contains
+  - assertion: file_contains
     path: "C:\\adp\\coreagent-only.log"
     pattern: "CORE"
     timeout: 60s
