+++ test/integration/cases/adp-memory-mode-disabled/config.yaml
@@ -15,14 +15,14 @@ env:
   DD_MEMORY_LIMIT: "1"
 
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: "Memory limiting disabled."
     timeout: 60s
   - parallel:
-      - type: process_stable_for
+      - assertion: process_stable_for
         duration: 10s
-      - type: log_not_contains
+      - assertion: log_not_contains
         pattern: "panic|PANIC"
         regex: true
         during: 10s
