+++ test/integration/cases/adp-rar-disabled/config.yaml
@@ -24,18 +24,18 @@ container:
   exposed_ports:
     - "58125/udp"
 
-assertions:
+procedure:
   # ADP should report registration failure (Core Agent returns error when RAR disabled). Allow 15s
   # (matching adp-cmd-port): the failure is only reported once the Core Agent's CMD/IPC API server is
   # reachable, which can take ~10s to come up on slower hosts.
-  - type: log_contains
+  - assertion: log_contains
     pattern: "Failed to register with the Datadog Agent. Registration will be retried periodically in the background."
     timeout: 15s
   # Make sure the process becomes healthy, and stays up without errors, for ~10 seconds after.
   - parallel:
-      - type: process_stable_for
+      - assertion: process_stable_for
         duration: 10s
-      - type: log_not_contains
+      - assertion: log_not_contains
         pattern: "panic|PANIC"
         regex: true
         during: 10s
