+++ test/integration/cases/dogstatsd-default-bind/config.yaml
@@ -22,14 +22,14 @@ env:
   DD_DATA_PLANE_DOGSTATSD_ENABLED: "true"
 
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: 'listen_addr:"udp://127.0.0.1:58125"'
     timeout: 60s
   - parallel:
-      - type: process_stable_for
+      - assertion: process_stable_for
         duration: 10s
-      - type: log_not_contains
+      - assertion: log_not_contains
         pattern: "panic|PANIC"
         regex: true
         during: 10s
