+++ test/integration/cases/dogstatsd-forwarding/config.yaml
@@ -28,50 +28,50 @@ container:
     - "58125/udp"
     - "9126/tcp"
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: "DogStatsD packet forwarding enabled"
     timeout: 15s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-udp-malformed"
     pattern: "this is not dogstatsd"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-udp-metric"
     pattern: "adp.forwarding.integration:1|c|#source:integration"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-stream-one"
     pattern: "adp.forwarding.stream_one:1|c"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-stream-two"
     pattern: "adp.forwarding.stream_two:2|c"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-uds-packet"
     pattern: "adp.forwarding.uds_datagram:1|c"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-forwarded-uds-stream-packet"
     pattern: "adp.forwarding.uds_stream:1|c"
     timeout: 30s
-  - type: file_contains
+  - assertion: file_contains
     path: "/tmp/dsd-metric-parsed"
     pattern: "component_events_received_total"
     timeout: 30s
   - parallel:
-      - type: process_stable_for
+      - assertion: process_stable_for
         duration: 10s
-      - type: port_listening
+      - assertion: port_listening
         port: 58125
         protocol: udp
         timeout: 10s
-      - type: port_listening
+      - assertion: port_listening
         port: 9126
         protocol: tcp
         timeout: 10s
-      - type: log_not_contains
+      - assertion: log_not_contains
         pattern: "Unsupported configuration key|panic|PANIC"
         regex: true
         during: 10s
