+++ test/integration/cases/dogstatsd-non-local-overrides-bind-host/config.yaml
@@ -29,14 +29,14 @@ env:
   DD_BIND_HOST: "10.9.8.7"
 
 
-assertions:
-  - type: log_contains
+procedure:
+  - assertion: log_contains
     pattern: 'listen_addr:"udp://0.0.0.0:58125"'
     timeout: 60s
   - parallel:
-      - type: process_stable_for
+      - assertion: process_stable_for
         duration: 10s
-      - type: log_not_contains
+      - assertion: log_not_contains
         pattern: "panic|PANIC"
         regex: true
         during: 10s
